Initial project setup
The project has been setup with a simple "Hello world!" output. This
is a TypeScript React project that creates a single JavaScript bundle.

ESLint and Prettier are used for linting, plus the ESLint `import`
plugin. The recommended ruless for ESLint, Prettier, the import plugin,
and the TypeScript ESLint plugn are all enabled. `tsc` is also used to
check types as part of the `lint` command.

AVA is the test runner, with `@testing-library/react` and Sinon also
being used for React testing and mocking respectively.

Rollup is used for bundling, with Babel being responsible for
transforming the TypeScript and transpiling.